src/main/java/org/promgrammers/springbootbasic/domain/customer/service/CustomerService.java
package org.promgrammers.springbootbasic.domain.customer.service;

import org.promgrammers.springbootbasic.domain.customer.dto.response.CustomerResponse;
import org.promgrammers.springbootbasic.domain.customer.dto.response.CustomersResponse;
import org.promgrammers.springbootbasic.domain.customer.model.Customer;
import org.promgrammers.springbootbasic.domain.customer.repository.CustomerRepository;
import org.springframework.stereotype.Service;

import java.util.List;
import java.util.stream.Collectors;

@Service
public class CustomerService {
    private final CustomerRepository customerRepository;

    public CustomerService(CustomerRepository customerRepository) {
        this.customerRepository = customerRepository;
    }

    public CustomersResponse findAll() {
        List<Customer> customers = customerRepository.findAll();
        return new CustomersResponse(customers.stream()
                .map(customer -> new CustomerResponse(customer.customerId(), customer.customerType()))
                .collect(Collectors.toList()));
    }
}
